refactor(src/components): refactor exampla and home
refactor the code, clean home and add example.tsx
