+++ src/components/DefaultDesignComponents/Inputs/Radio/index.tsx
@@ -1,0 +1,39 @@
+import type { NextPage } from 'next';
+import { ReactNode } from 'react';
+import { RadioButton, RadioFill, RadioRoot } from './styles';
+
+// COMPONENTE EXEMPLO PARA O RADIO BUTTON LOGIC
+interface RadioProps {
+  name: string;
+  value: string;
+  checked: boolean;
+  onChange: (event: any) => void | Function;
+  children: ReactNode;
+}
+
+const Radio: NextPage = ({
+  name,
+  value,
+  checked,
+  onChange,
+  children
+}: RadioProps) => (
+  <div>
+    <RadioRoot>
+      <label htmlFor={name}>
+        {children}
+        <RadioButton
+          type="radio"
+          onChange={onChange}
+          name={name}
+          value={value}
+          checked={checked}
+          aria-checked={checked}
+        />
+        <RadioFill />
+      </label>
+    </RadioRoot>
+  </div>
+);
+
+export default Radio;
